animation: Allow spinner to be sizes other than 16x16 pixels
Since sliced images are provided in a two dimensional grid, instead
of linear strip, there's no way to infer the size of an individual
cell inside a sliced image. Because of this limitation the texture
cache api for loading an image requires the cell size (called grid
size in the api) be explicitly specified. But because sliced images
are scalable SVGs, the texture cache code needs to know the output
animation size as well, to ensure the animation is rendered with an
appropriately high resolution. The API doesn't ask for this
information, however, instead just assuming the output size is the same
as the cell size. That effectively means that spinners can only be
16x16 pixels big (because our spinner slicable images have cells that
are nominally 16x16).

This commit changes the API to allow the output image size to be
specified as well, and fixes the code to render at the appropriate
output size.