unlockDialog: Add menus for auth mechanisms
Now that we have a list of auth mechanisms available, we need
a way to select one from the unlock screen.

This commit adds a new Unlock Options menu, that matches the
same style as the one on the Login Options menu on the login
screen, speced out in the design here:

https://gitlab.gnome.org/Teams/Design/os-mockups/-/blob/master/lock-login/login-authentication.png

This commit also adds a menu of indicators for background authentication
mechanisms.